[QA13814]
feat(QA13814)
Costi notifica fase 5.
Aggiunta scenari per modalità di pagamento ASYNC.
Riorganizzazione degli scenari in più file feature per migliore leggibilità
Aggiunto metodo di utils per disabilitare i controlli sugli audit log in fase di debugging.
Edit del client B2b per chiamata all'api di recupero costi di delivery push;
Creazione della classe per eseguire la suite di test;
Edit del pom con l'open-api definito mergiato su develop

+++ src/test/java/it/pagopa/pn/cucumber/steps/pa/utilityVersions/AwsUtils.java
@@ -11,7 +11,6 @@ import software.amazon.awssdk.services.cloudwatchlogs.model.FilterLogEventsReque
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
 import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
 import software.amazon.awssdk.services.dynamodb.model.QueryRequest;
-import software.amazon.awssdk.services.dynamodb.model.QueryResponse;
 
 import java.util.Map;
 
@@ -77,8 +76,8 @@ public class AwsUtils {
                 .build();
     }
 
-    public static FilterLogEventsRequest buildCloudWatchLogRequest(String microservice, String search, int minutes) {
-        long startTime = System.currentTimeMillis() - (minutes * 60 * 1000);
+    public static FilterLogEventsRequest buildCloudWatchLogRequest(String microservice, String search, long minutes) {
+        long startTime = System.currentTimeMillis() - (minutes * 60L * 1000L);
         return FilterLogEventsRequest.builder()
                 .logGroupName(microservice)
                 .filterPattern(search)
